+++ pdclient/src/lib/functions.js
@@ -1,8 +1,8 @@
 
 "use client"
 import {  useState, useEffect } from 'react'
-import Link from 'next/link'
-import { createUser } from "@/app/actions"
+
+import { createUser } from "@/lib/actions"
 
 
 export function DisplayUsers(){
@@ -72,20 +72,5 @@ export function AddUser() {
 }
 
 
-export function NavBar(){
-  return(
-    <div className="navbar bg-base-100">
-      <div className="navbar-start">
-        <Link href='/' className="btn btn-ghost normal-case text-3xl text-orange-500">PodRace</Link>
-      </div>
-      <div className="navbar-end">
-            <Link href="/create">
-            <button className="btn btn-wide bg-orange-500 text-white">Join the Race</button>
-            </Link>
-      </div>
-    </div>
-
-  )
-}
